handle input errors

+++ mspr2_back_springboot_api/src/main/java/com/capus/securedapi/controllers/ConcertController.java
@@ -3,11 +3,15 @@ package com.capus.securedapi.controllers;
 
 
 import java.util.List;
-import java.util.Map;
 
+import com.capus.securedapi.dto.InformationDto;
+import com.capus.securedapi.payload.response.MessageResponse;
+import com.capus.securedapi.repository.ConcertRepository;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.Parameters;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -24,43 +28,64 @@ import com.capus.securedapi.service.ConcertService;
 public class ConcertController {
 	
 	private final ConcertService concertService;
+	private final ConcertRepository concertRepository;
 
-	public ConcertController(ConcertService concertService) {
+	public ConcertController(ConcertService concertService, ConcertRepository concertRepository) {
 		this.concertService = concertService;
+		this.concertRepository = concertRepository;
 	}
-	
-	//ADD REST API POST add account
+
+	@ApiResponses({
+			@ApiResponse(responseCode = "201", content = { @Content(mediaType = "application/json",
+					schema = @Schema(implementation = ConcertDto.class)) })
+	})
 	@Operation(
 			summary = "Create a new concert",
 			description = "ONLY FOR EDITORS.",
 			tags = { "Editor only" })
 	@PostMapping
-	@PreAuthorize("hasRole('EDITOR') or hasRole('ADMIN')")
+	@PreAuthorize("hasRole('EDITOR')")
 	public ResponseEntity<ConcertDto> addConcert(@RequestBody ConcertDto concertDto){
-		System.out.println(concertDto);
 		return new ResponseEntity<>(concertService.createConcert(concertDto), HttpStatus.CREATED);
 	}
-	
-	//ADD REST API GET account get By Id
+
+	@ApiResponses({
+			@ApiResponse(responseCode = "200", content = { @Content(mediaType = "application/json",
+					schema = @Schema(implementation = ConcertDto.class)) }),
+			@ApiResponse(responseCode = "400", description = "No Concert found with id: +id")
+	})
 	@Operation(
 			summary = "Get a concert by ID",
 			description = "All access allowed.",
 			tags = { "All access allowed" })
 	@GetMapping("{id}")
-	public ResponseEntity<ConcertDto> getConcertById(@PathVariable Long id){
-		System.out.println("Get Concert by ID request");
+	public ResponseEntity<?> getConcertById(@PathVariable Long id) {
+		if (!concertRepository.existsById(id)) {
+			return ResponseEntity
+					.internalServerError()
+					.body(new MessageResponse("No concert found with id: "+id));
+		}
 		ConcertDto concertDto = concertService.getConcertById(id);
 		return ResponseEntity.ok(concertDto);
 	}
-	
-	// ADD REST API POST update
+
+	@ApiResponses({
+			@ApiResponse(responseCode = "200", content = { @Content(mediaType = "application/json",
+					schema = @Schema(implementation = ConcertDto.class)) }),
+			@ApiResponse(responseCode = "400", description = "No Concert found with id: +id")
+	})
 	@Operation(
 			summary = "Update a concert by ID",
 			description = "ONLY FOR EDITORS.",
 			tags = { "Editor only" })
 	@PutMapping("update/{id}")
-	@PreAuthorize("hasRole('EDITOR') or hasRole('ADMIN')")
-	public ResponseEntity<ConcertDto> update(@PathVariable Long id,@RequestBody ConcertDto request){
+	@PreAuthorize("hasRole('EDITOR')")
+	public ResponseEntity<?> update(@PathVariable Long id,@RequestBody ConcertDto request){
+		if (!concertRepository.existsById(id)) {
+			return ResponseEntity
+					.internalServerError()
+					.body(new MessageResponse("No concert found with id: "+id));
+		}
 		ConcertDto updatedConcertDto = concertService.update(id,request.getNom(),request.getImage(), request.getDescription(), request.getOrigine(), request.getDate(), request.getHeure(), request.getScene(), request.getLien());
 		return ResponseEntity.ok(updatedConcertDto);
 		
@@ -79,14 +104,22 @@ public class ConcertController {
 		return ResponseEntity.ok(concertDtos);
 	}
 
-	//DELETE ACCOUNT BY ID REST API
+	@ApiResponses({
+			@ApiResponse(responseCode = "200", description = "Deleted Concert "),
+			@ApiResponse(responseCode = "400", description = "No Concert found with id: +id")
+	})
 	@Operation(
 			summary = "Delete a concert by ID",
 			description = "ONLY FOR EDITORS.",
 			tags = { "Editor only" })
 	@DeleteMapping("{id}")
-	@PreAuthorize("hasRole('EDITOR') or hasRole('ADMIN')")
-	public ResponseEntity<String> deleteConcert(@PathVariable Long id){
+	@PreAuthorize("hasRole('EDITOR')")
+	public ResponseEntity<?> deleteConcert(@PathVariable Long id){
+		if (!concertRepository.existsById(id)) {
+			return ResponseEntity
+					.internalServerError()
+					.body(new MessageResponse("No concert found with id: "+id));
+		}
 		System.out.println("Delete by ID request");
 		concertService.deleteConcert(id);
 		return ResponseEntity.ok(id+" Deleted");
