+++ mspr2_back_springboot_api/src/main/java/com/capus/securedapi/controllers/InformationController.java
@@ -1,8 +1,15 @@
 package com.capus.securedapi.controllers;
 
 import com.capus.securedapi.dto.InformationDto;
+import com.capus.securedapi.dto.PointeurDto;
+import com.capus.securedapi.payload.response.MessageResponse;
+import com.capus.securedapi.repository.InformationRepository;
 import com.capus.securedapi.service.InformationService;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -18,18 +25,23 @@ import java.util.List;
 public class InformationController {
 
     private final InformationService informationService;
+    private final InformationRepository informationRepository;
 
-    public InformationController(InformationService informationService) {
+    public InformationController(InformationService informationService, InformationRepository informationRepository) {
         this.informationService = informationService;
+        this.informationRepository = informationRepository;
     }
 
-    //ADD REST API POST add account
+    @ApiResponses({
+            @ApiResponse(responseCode = "201", content = { @Content(mediaType = "application/json",
+                    schema = @Schema(implementation = InformationDto.class)) })
+    })
     @Operation(
             summary = "Create a new info",
             description = "ONLY FOR EDITORS.",
             tags = { "Editor only" })
     @PostMapping
-    @PreAuthorize("hasRole('EDITOR') or hasRole('ADMIN')")
+    @PreAuthorize("hasRole('EDITOR')")
     public ResponseEntity<InformationDto> addInformation(@RequestBody InformationDto informationDto) {
         return new ResponseEntity<>(informationService.createInformation(informationDto), HttpStatus.CREATED);
     }
@@ -46,26 +58,43 @@ public class InformationController {
         return ResponseEntity.ok(informationDtos);
     }
 
-    // ADD REST API POST update
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", content = { @Content(mediaType = "application/json",
+                    schema = @Schema(implementation = InformationDto.class)) }),
+            @ApiResponse(responseCode = "400", description = "No Information found with id: +id")
+    })
     @Operation(
             summary = "Update an info by ID",
             description = "ONLY FOR EDITORS.",
             tags = { "Editor only" })
     @PutMapping("update/{id}")
-    @PreAuthorize("hasRole('EDITOR') or hasRole('ADMIN')")
-    public ResponseEntity<InformationDto> update(@PathVariable Long id, @RequestBody InformationDto request) {
+    @PreAuthorize("hasRole('EDITOR')")
+    public ResponseEntity<?> update(@PathVariable Long id, @RequestBody InformationDto request) {
+        if (!informationRepository.existsById(id)) {
+            return ResponseEntity
+                    .badRequest()
+                    .body(new MessageResponse("No Information found with id: "+id));
+        }
         InformationDto updatedInformationDto = informationService.update(id, request.getMessage(), request.isImportant());
         return ResponseEntity.ok(updatedInformationDto);
     }
 
-    //DELETE ACCOUNT BY ID REST API
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "Deleted Information "),
+            @ApiResponse(responseCode = "400", description = "No Information found with id: +id")
+    })
     @Operation(
             summary = "Delete an info by ID",
             description = "ONLY FOR EDITORS.",
             tags = { "Editor only" })
     @DeleteMapping("{id}")
     @PreAuthorize("hasRole('EDITOR') or hasRole('ADMIN')")
-    public ResponseEntity<String> deleteInformation(@PathVariable Long id) {
+    public ResponseEntity<?> deleteInformation(@PathVariable Long id) {
+        if (!informationRepository.existsById(id)) {
+            return ResponseEntity
+                    .badRequest()
+                    .body(new MessageResponse("No Information found with id: "+id));
+        }
         informationService.deleteInformation(id);
         return ResponseEntity.ok("Deleted Information " + id);
     }
